fixed lclean permission with external.conf file

+++ bin/python/rm_interface.py
@@ -4,6 +4,8 @@ import sys
 import lstart_create as lc
 import utils as u
 import pwd
+from sys import platform as _platform
+from netkit_commons import LINUX, LINUX2
 
 parser = argparse.ArgumentParser()
 parser.add_argument('path')
@@ -18,19 +20,24 @@ if args.directory:
 #delete subinterface created with vlan_id
 def remove_subinterface(path):
 	delete_commands = []
-	temp_path = os.path.join(pwd.getpwuid(os.getuid()).pw_dir, "netkit_temp/")
-	hash_path = str(u.generate_urlsafe_hash(path))
-	if (os.path.exists(temp_path + hash_path + '_external_links')):
-		if (os.stat(temp_path + hash_path + '_external_links').st_size != 0):
-			with open(temp_path + hash_path + '_external_links', 'r') as external_file_temp:
-				delete_commands.append('echo "\033[0;33mSubinterfaces will be deleted\033[0m"')
-				for line in external_file_temp:
-					subinterface = line.split()[0]
-					delete_commands.append('sudo ip link set dev ' + subinterface + ' ' + 'down')
-					delete_commands.append('sudo ip link delete ' + subinterface)
-					delete_commands.append('echo ' + subinterface)
+	if (_platform == LINUX or _platform == LINUX2):
+		if (os.geteuid() == 0):
+			temp_path = os.path.join(pwd.getpwuid(os.getuid()).pw_dir, "netkit_temp/")
+			hash_path = str(u.generate_urlsafe_hash(path))
+			if (os.path.exists(temp_path + hash_path + '_external_links')):
+				if (os.stat(temp_path + hash_path + '_external_links').st_size != 0):
+					with open(temp_path + hash_path + '_external_links', 'r') as external_file_temp:
+						delete_commands.append('echo "\033[0;33mSubinterfaces will be deleted\033[0m"')
+						for line in external_file_temp:
+							subinterface = line.split()[0]
+							delete_commands.append('sudo ip link set dev ' + subinterface + ' ' + 'down')
+							delete_commands.append('sudo ip link delete ' + subinterface)
+							delete_commands.append('echo ' + subinterface)
 
-		os.remove(temp_path + hash_path + '_external_links')
+				os.remove(temp_path + hash_path + '_external_links')
+		else:
+			sys.stderr.write("Please you started the lab with external.conf file, need root permission to clean current lab." +"\n")
+			sys.exit(1)
 
 	return delete_commands
